Changes made to correct problem with nakes of files in data directory.

+++ Python_Random_Forest/src/UtilRecords.py
@@ -26,9 +26,11 @@ def read_from_csv(input_file):
     # Note that we must specify the right type of encoding in order to read in all characters
     # correctly.  Some of the data contain Greek letters which me must account for.
     '''
+    if input_file.startswith("data"):
+        input_file = "..\\" + input_file
     
     if not Path(input_file).exists():
-        input_file = "..\\" + input_file
+        AssertionError("input_file not found")
     
     df = pd.read_csv(input_file, skip_blank_lines = False, 
                      na_filter = True, low_memory = False,
@@ -53,7 +55,7 @@ def write_to_csv(df, file_output):
     file_output: text
         output file name
     '''
-    if not Path(file_output).exists():
+    if file_output.startswith("data"):
         file_output = "..\\" + file_output
         
     # Write DataFrame to output.
@@ -79,8 +81,9 @@ def remove(filename):
     # remove file
     if Path(filename).exists(): os.remove(filename)
     else:
-        filename = "..\\" + filename
-        if Path(filename).exists(): os.remove(filename)
+        if filename.startswith("data"):
+            filename = "..\\" + filename
+            if Path(filename).exists(): os.remove(filename)
     
 
 def replace_null_with_none(df):
